Split off fields.
